Changes setSource to preserve selection.
This function is used in toDataSource bindings. This fixes a problem when you also have a binding to a ListControl's selected value or item and you want to change or update the data source but preserve the selection. When a ListControl's DataSource changes, the selection is reset. If the model's selection property hasn't changed, the framework does not trigger the model-to-view binding to set the selection. Not only that, but the view-to-model binding for the selection does not fire either even though the selection was reset, so the form and the model would get out of sync. This change makes the update to the DataSource preserve the selection when possible so that bindings to the selection will behave as expected.