Removes thread context switching for async handlers.
I originally added this because I figured that _everything_ in an async handler should be executed on a background thread automatically, because why else would you want an async handler if not to run things in the background so the UI can remain responsive? I have since learned better how Async works or should work. It seems there are two main use cases for Async workflows: CPU-intensive work and IO work.

If you have Async IO work, you can simply await it from your GUI thread and it won't be blocked because the task/async is sleeping while waiting for IO, no need to switch threads.

If you have CPU-intensive work, you can switch contexts yourself in your handler, like the Framework was doing previously, to make sure the work happens on a different thread and not block the GUI thread. The other option is to call Async.StartChild to delegate the work to another thread.

Either way, it is presumptuous of the framework to execute async handlers on a background thread and probably a bad design choice because:
- it adds overhead that may not be necessary at all
- would be surprising behavior to users not familiar with the inner workings of the framework
- is problematic if you need to make changes to the GUI or open a new GUI from an async handler

This was a good resource: http://tomasp.net/blog/async-non-blocking-gui.aspx/