Fixes potential infinite loop with two-way bindings.
If a "control" with a setter that always fires PropertyChanged is bound two-way, an infinite loop would occur. Most actual controls check that the property actually changed before firing the event, so this wasn't a big problem, however the framework shouldn't be causing this kind of feedback loop. This fix makes sure that it doesn't trigger two-way bindings on the same property when updating the view in response to view changed.